Added a new file with a cleaner implementation of defcustom
defcustom isnt implemented here yet, but this implementation is more
broken up and easier to grok. it should be easier to implement as
well. this new attempt tries to solve the problem of validating values
when the place is not a symbol to be dispatched upon but instead an
accessor or function with a setf expander already defined.