Refining atomic setv (#4)
* added setv-atomic macro for resetting changes on error

* updated examples

* Reworked searching for objects, added predicate

* added descriptions of more exported symbols

* removed some exported symbols

* rewrote remove-keys, destructuring-keys, setv, and setv-atomic

reworked remove-keys to allow parsing of keyargs with default
values. added destructuring-keys, a variation on destructuring bind,
which splits the keyargs from everything else. rewrote setv to use
destructuring-keys.

* add trivial-cltl2

* implementing with-atomic-setv using compiler-let.

* implemented basic with-atomic-setv

Its working now, but we are currently resetting to the default value,
NOT the previous value as we should. This needs to be handled in
%atomic-setv-reset

* Begining to refine with-atomic-setv

currently not working, and complaining that place value and db arent
variables, but they should be...

* fixed up reset-place

allow for resetting to previous-value instead of default-value

* fixed double qquote, and double push

at this point, with-atomic-setv should be working as intended, but
more testing is needed.

* finalized sytnax shape to allow for further customization

* added with-atomic-setv examples.

* reformatted

* rewrote %setv, modified with-atomic-setv

%setv was rewritten - broke out several parts for readability. this
could use more work, to really break out the portions that could use
it. in other words we made a mess and we need to clean up.

with-atomic-setv was modified to re-error if we error out of the
block, and if not we return the return value from body.

* added missing condition

* exported setv-atomic and with-atomic-setv

* handle any errors in body of with-atomic-setv

reshaped the structure of the places being passed around at
macroexpansion time to track the database in question in order to
handle errors occuring outside of a setv form.

* documented with-atomic-setv

* fixed global reset for with-atomic-setv, changed accessor

* exported reset-place

* added database keyarg to defconfig

* implemented tests using fiveam