added macros, formatting, buttons, and more
added the macro monospaced, which wraps with-text-family with the
:fix family selected.

added change-layout macro which takes a frame and a layout and checks
if that layout is already set before changing to it.

added reinitialize and regen-config to the defconfig calls.

added current-config-info slot to the application frame

added config-info-view pane and layout for it

added app-main function to ease running the app

changed com-refresh-frame to be named "Refresh"

added presentation type, command, and command-translator to switch to
the config-info oriented layout.

modified write-database-to-stream to show information about the
symbol/accessor in a better manner.

added functions to display a config info object. these should include
buttons to reset values, but they dont refresh the frame afterwards
despite calls to com-refresh-frame.