added ability to specify which errors to handle in with-atomic-setv
added the keyarg handle-errors to with-atomic-setv, which allows one to only
handle specific error types and let others propogate up. this is useful if one
wants to only reset on invalid values, and not reset on other errors outside of
a setv form.

rewrote setv-wrapped-error to descend from config-error instead of error, and
added documentation to it.

added light documentation to helper macros and documented %setv. this
documentation should probably be moved to setv.