move with-atomic-setv computations to runtime
I'm not sure why i thought compiler-let was neccessary - its not the most
obvious way to do this. reimplemented all with-atomic-setv helper macros to do
their computations of what to reset at runtime instead of macroexpansion time,
and changed the names to compiler-time/runtime depending on whic they
are. reimplemented with-atomic-setv to just call with-runtime-atomic-setv. work
can be done on moving with-atomic-setv back to using compiler-let (i think it
will be more performant to do the computation at macroexpansion time, as it will
cut down on looping through the list multiple times at runtime) at a later
point. the important thing right now is that with-atomic-setv now works as
intended (i hope). this should be added to the test suite. the issue was:

(setv *bounded-number* 0)
(with-new-atomic-setv ()
  (setv *bounded-number* 1)
  (setv *bounded-number* 2)
  (setv *bounded-number* 50)) ; 50 is out of range

would end with *bounded-number* not being zero in the macroexpansion time
version, because we used reset-place instead of tracking their values before
setting them. in the above example this mean that prev was 0 and current was 0,
then prev was 0 and current was 1, then prev was 1 and current was 2, then we
hit an invalid value and reset current to prev, which was 1, not 0 as we
wanted.