moved to generics, rewrote with-config-info
rewrote with-config-info to take advantage of the new function
%fsetv-get-config-info-object, which handles errors upon not finding a config
info object in the database provided. the function %%with-config-info is the new
expansion, and it installs restarts in case the policy is :strict - the user can
use supply a new database.

moved reset-computed-place and clean-previous-value to generic functions, with
four methods each, which allow for symbols (ie regular config-info objects)
lists (ie accessor-config-info objects) and the objects themselves to be
supplied.