move towards separate param/accessor defconfig macros
Broke out classes into config-info and accessor-config-info, and began
formalizing the differences between the two. namely accessors cant track
previous or default values.

added conditions for untrackable places - no-bound-default-value-error is
signalled when the default-value slot is unbound, and not-resettable-place-error
is signalled when one tries to set a resettable place.

wrote additional macros for defining defconfig objects, in an attempt to
separate accessor definitions and parameter/var definitions. next step is to
move things into functions.