added *setv-permissiveness*, rewrote setv, added setv helpers
this is part of a major move towards installing restarts and handlers for setv
expansions, but also for moving things into functions as much as possible.

defined *setv-permissiveness*, which governs how setv should handle things when
A) a config-info object isnt found, and B) when a value is invalid.

added helper macro %setv-ensure-setf which handles the minutae of setting a
place to a value.

modified %setv-coerced, but it isnt used anymore.

modified %setv-original, but it isnt used anymore.

added helper macro %setv-no-config-error, but it isnt used.

modified %setv, but it isnt used anymore.

added %fsetv-ensure-validity, which returns either a valid value, a symbol
indicating invalidity, or errors out with an invalid-datum-error or a
invalid-coerced-datum-error. Restarts are installed around the errors to provide
a value and optionally have it validated, or set regardless. this function is
called only by %%setv and %%setv-coerced.

added %fsetv-invalid-value, but it isnt used.

added %%setv-coerced, which ensures validity and sets to the propper value, or
it signals an error.

added %fsetv-get-config-info-object, which tries to get a config-info object out
of a database, or signals an error. it installs a handler which is dependent
on *setv-permissiveness* regarding what it does.

added macro %%setv, which replaces %setv, and handles getting a config-info
object and using it to validate a value, or further expand via
%%setv-coerced.

modified setv to expand using %%setv instead of %setv.