changed failing test to work
we want to distinguish from w-a-s and w-a-s*. w-a-s is simpler, and
has slightly less complicated semantics, but ONLY works for dynamic
variables, while w-a-s* is more complexe, more difficult to debug (due
to the compiler let, which sly doesnt seem to know how to deal with
during macro expansion), but handles both dynamic variables and
accessors. Until we have determined how to reconcile their differences
the test in question SHOULD equal 8, to indicate that the slot was
modified despite being within a call to w-a-s.