updated defconfig definition
added the helper macros defconf-v and defconf-a, which define a
configuration object of the appropriate type. defconf-v signals an
error if the default value provided is invalid.

added exported macros define-accessor-config and
define-variable-config, which wrap around defconf-a/v but generate a
predicate function out of the typespec when its provided.

rewrote defconfig to use these new macros, and take a &rest argument
instead of a set of &key arguments, which makes it easier to
generalize it for accessors and variables. If it generates a variable
config it will also define a dynamic variable using either
defparameter or defvar. A restart is in place to ignore a failure and
generate a variable definition regarless of a missing config.