reworked runtime setv to handle accessor reset
with-atomic-setv (no star⟅*⟆) now supports resetting accessors. Uff
was this an odd one. I was overthinking this to an incredible degree;
I wanted to have a lexical equivalent of symbol-value, but i had
neglected to remember that I could just store a reference to a lexical
variable as we are doing things at macroexpansion time and can splice
in the variable itself. So, now we track the (quoted) variable/place,
the value, and the variable reference. This is a bit simplistic and it
should be noted that any side effects of calling an accessor will be
run TWICE. For example, (accessor (progn (print 'hi) *v*)) will print
'hi once upon setting, and once again upon pushing to the tracking
list. this is different to with-atomic-setv* which will only cause
these side effects twice if were resetting. This should be
standardized, and should be considered a bug.

We also added (another) gensym to prevent variable capture of the
accumulator, removed some comments and a useless copy of
with-runtime-atomic-setv, renamed the test suite package, fixed a
failing test (which was failing because the behavior is now
correct...), and changed some formatting.