Update README formatting, add metaclass example
Update the readme file to not depend on a visual-line-mode or similar line
wrapping functionality being present.

Add a section comparing using accessor configs vs using a metaclass to enforce a
type upon a location. the gist of it is that metaclass is good for slots,
accessor configs are good for general setf-able functions.